+++ sdk/csharp/Partio.Sdk/Models/CompletionEndpoint.cs
@@ -77,9 +77,9 @@ namespace Partio.Sdk.Models
         public Dictionary<string, string> Tags { get; set; } = new Dictionary<string, string>();
 
         [JsonPropertyName("CreatedUtc")]
-        public DateTime CreatedUtc { get; set; }
+        public DateTime CreatedUtc { get; set; } = DateTime.UtcNow;
 
         [JsonPropertyName("LastUpdateUtc")]
-        public DateTime LastUpdateUtc { get; set; }
+        public DateTime LastUpdateUtc { get; set; } = DateTime.UtcNow;
     }
 }