+++ sdk/csharp/Partio.Sdk/Models/UserMaster.cs
@@ -35,10 +35,10 @@ namespace Partio.Sdk.Models
         public Dictionary<string, string>? Tags { get; set; }
 
         [JsonPropertyName("CreatedUtc")]
-        public DateTime CreatedUtc { get; set; }
+        public DateTime CreatedUtc { get; set; } = DateTime.UtcNow;
 
         [JsonPropertyName("LastUpdateUtc")]
-        public DateTime LastUpdateUtc { get; set; }
+        public DateTime LastUpdateUtc { get; set; } = DateTime.UtcNow;
 
         [JsonPropertyName("Password")]
         public string? Password { get; set; }